+++ src/main/java/edu/hniu/imchatroom/model/bean/FriendShip.java
@@ -17,7 +17,7 @@ public class FriendShip {
 
     private static final long serialVersionUID = 5362452342353454L;
     @JsonProperty("fsId")
-    private Integer fsId;
+    private Long fsId;
     /*private Integer hostId;
     private Integer friendId;*/
     private User hostUser;
